@@ -23,6 +23,7 @@
     "chart.js": "^2.7.3",
     "chokidar": "^3.4.3",
     "classnames": "^2.2.6",
+    "clean-webpack-plugin": "^4.0.0",
     "connect-history-api-fallback": "^1.6.0",
     "cookie-parser": "^1.4.5",
     "dart-sass": "^1.25.0",
@@ -33,6 +34,7 @@
     "express": "^4.16.3",
     "filesize": "^4.1.2",
     "glob": "^9.3.4",
+    "html-webpack-plugin": "^5.5.1",
     "iconv-lite": "^0.6.3",
     "ini": "^2.0.0",
     "internal-ip": "^6.0.0",
@@ -52,15 +54,15 @@
     "qrcode-terminal": "^0.12.0",
     "qrcode.react": "^3.1.0",
     "query-string": "^6.2.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-chartjs-2": "^2.7.4",
     "react-click-outside": "^3.0.1",
     "react-collapse": "^5.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-dplayer": "^0.4.2",
     "react-modal": "^3.12.1",
     "react-router-dom": "^4.3.1",
-    "react-toastify": "^6.0.5",
+    "react-toastify": "^9.1.2",
     "react-visibility-sensor": "^5.1.1",
     "request": "^2.88.2",
     "request-promise": "^4.2.6",
@@ -71,6 +73,9 @@
     "sweetalert2": "^8.0.7",
     "trash": "^6.1.1",
     "underscore": "^1.9.1",
+    "webpack": "^5.82.1",
+    "webpack-cli": "^5.1.1",
+    "webpack-dev-server": "^4.15.0",
     "whatwg-fetch": "^3.0.0",
     "winston": "^3.2.1"
   },
@@ -82,7 +87,7 @@
     "babel-eslint": "^10.0.0",
     "babel-loader": "^8.0.0",
     "chai": "^4.2.0",
-    "clean-webpack-plugin": "^1.0.0",
+    "clean-webpack-plugin": "^4.0.0",
     "concurrently": "^4.0.0",
     "css-loader": "^2.1.0",
     "eslint": "^5.0.0",
@@ -92,7 +97,7 @@
     "eslint-plugin-react": "^7.7.0",
     "file-loader": "^3.0.0",
     "flow-bin": "^0.91.0",
-    "html-webpack-plugin": "^3.2.0",
+    "html-webpack-plugin": "^5.5.1",
     "less": "^3.9.0",
     "less-loader": "^4.1.0",
     "mocha": "^5.2.0",
@@ -100,8 +105,8 @@
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.1",
     "url-loader": "^1.0.1",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.0"
+    "webpack": "^5.82.1",
+    "webpack-cli": "^5.1.1",
+    "webpack-dev-server": "^4.15.0"
   }
 }
