@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
 import './style/bootstrap.min.css';
 
-ReactDOM.render( <BrowserRouter>
+createRoot(document.getElementById('root')).render(<BrowserRouter>
                         <App />
-                        </BrowserRouter>
-                , document.getElementById('root'));
+                        </BrowserRouter>);
